Logisticsgpt/requirements.txt
llama_index>=0.10.40
llama-cpp-python>=0.2.67
qdrant-client>=1.8.0
datasets>=2.18.0
tqdm>=4.66.0
langchain>=0.1.20
langchain-openai>=0.1.7
sqlalchemy>=2.0
psycopg2-binary>=2.9
matplotlib>=3.8
pandas>=2.2
